refactoring around CLI
