+++ cmd/lsec2/version/version.go
@@ -1,0 +1,4 @@
+package version
+
+// VERSION says my version number
+const VERSION = "0.3.0"
