+++ scripts/install.sh
@@ -4,10 +4,6 @@ set -ex
 
 source scripts/_prepare.sh
 
-[ $# -lt 1 ] && { echo 'need to assign output target'; exit 1; }
-OUTPUT=${1}
-shift
-
 OTHER_OPTS=""
 if [ $# -ge 1 ]
 then
@@ -18,4 +14,4 @@ fi
 LDFLAGS=$(${MYDIR}/_ldflags.sh)
 
 echo "LDFLAGS=${LDFLAGS}, GOOS=${GOOS}, GOARCH=${GOARCH}, OTHER_OPTS=${OTHER_OPTS}"
-go build -a -tags netgo -installsuffix netgo -ldflags="${LDFLAGS}" ${OTHER_OPTS} -o ${OUTPUT} ${CMD_GO_DIR}
+go install -v -ldflags="${LDFLAGS}" ${OTHER_OPTS} ${CMD_GO_DIR}
