--- version.go
@@ -1,4 +1,0 @@
-package main
-
-// VERSION info
-const VERSION = "0.3.0"
